Break app modals into their own components
