+++ src/components/App.tsx
@@ -1,28 +1,18 @@
-import React, { useContext, useState, useEffect } from "react"
+import React, { useContext } from "react"
 import styled from "styled-components"
 
 import {
   navigationStateContext,
   navigationDispatchContext,
 } from "reducer-contexts/navigation"
-import {
-  polygonGroupsDispatchContext,
-  polygonGroupsStateContext,
-} from "reducer-contexts/polygon-groups"
-import {
-  backgroundDispatchContext,
-  backgroundStateContext,
-} from "reducer-contexts/background"
+import { polygonGroupsDispatchContext } from "reducer-contexts/polygon-groups"
 
 import Screens from "components/Screens"
 import { MainCanvas } from "components/MainCanvas"
 import { GroupsDisplay } from "components/GroupsDisplay"
 import { PolygonDisplay } from "components/PolygonDisplay"
-import { ModalBox } from "components/ModalBox"
-import { ToggleSwitch } from "components/ToggleSwitch"
-import { Slider } from "components/Slider"
-import { P5Canvas } from "./P5Canvas"
-import { generateGifSketch } from "polygon-logic/polygon-p5-draw"
+import { GenerateGifModal } from "./ShareGifModal"
+import { EditBackgroundModal } from "./EditBackgroundModal"
 
 const Main = styled.main`
   display: grid;
@@ -48,183 +38,11 @@ const Navigation = styled.nav`
   box-shadow: inset 0px 4px 9px -2px #404040;
 `
 
-const NavigationButton = styled.button`
+export const NavigationButton = styled.button`
   min-height: 50px;
   border-radius: 5px;
   margin: 5px;
 `
-
-const ColourPickerInput = styled.input`
-  box-sizing: border-box;
-  width: 100%;
-  height: 35px;
-  padding: 5px;
-  margin: 5px 0 0;
-  border-radius: 5px;
-`
-
-const ModalInternalWrappingDiv = styled.div`
-  display: grid;
-  grid-gap: 10px;
-  font-size: 20px;
-`
-
-const ModalUpdateButton = styled.button`
-  display: block;
-  width: 80%;
-  height: 50px;
-  margin: 10px auto 0;
-  font-size: 18px;
-  border-radius: 5px;
-`
-
-const EditBackgroundModal: React.FC = () => {
-  const backgroundDispatch = useContext(backgroundDispatchContext)
-  const backgroundState = useContext(backgroundStateContext)
-
-  const [canUpdate, setCanUpdate] = useState(false)
-  const [editModalIsClosed, setEditModalIsClosed] = useState(true)
-  const [shouldRedrawBackground, setShouldRedrawBackground] = useState(
-    backgroundState.shouldRedraw
-  )
-  const [backgroundHex, setBackgroundHex] = useState(backgroundState.hex)
-  const [backgroundOpacity, setBackgroundOpacity] = useState(
-    backgroundState.opacity
-  )
-
-  useEffect(() => {
-    if (
-      shouldRedrawBackground !== backgroundState.shouldRedraw ||
-      backgroundHex !== backgroundState.hex ||
-      backgroundOpacity !== backgroundState.opacity
-    ) {
-    }
-    setCanUpdate(true)
-  }, [
-    backgroundState,
-    shouldRedrawBackground,
-    backgroundHex,
-    backgroundOpacity,
-  ])
-
-  return (
-    <ModalBox
-      buttonText="Edit Background"
-      title="Edit Background"
-      StyledButton={NavigationButton}
-      isClosed={editModalIsClosed}
-      setIsClosed={setEditModalIsClosed}
-    >
-      <ModalInternalWrappingDiv>
-        <ToggleSwitch
-          checked={shouldRedrawBackground}
-          id="redraw-background-toggle"
-          label="Redraw Background"
-          setFunction={setShouldRedrawBackground}
-          checkedText={{ checked: "Yes", unchecked: "No" }}
-        />
-        <div>
-          <label htmlFor="background-colour-picker">Colour</label>
-          <ColourPickerInput
-            id="background-colour-picker"
-            type="color"
-            value={backgroundHex}
-            onChange={({ currentTarget: { value } }) => {
-              setBackgroundHex(value)
-            }}
-          />
-        </div>
-        <Slider
-          label="Opacity"
-          id="background-opacity"
-          currentValue={backgroundOpacity}
-          max={100}
-          min={0}
-          setFunction={setBackgroundOpacity}
-        />
-        <ModalUpdateButton
-          type="button"
-          disabled={!canUpdate}
-          onClick={() => {
-            backgroundDispatch({
-              type: "UPDATE_BACKGROUND_WITH_HEX",
-              shouldRedraw: shouldRedrawBackground,
-              hexColour: backgroundHex,
-              opacity: backgroundOpacity,
-            })
-            setCanUpdate(false)
-            setEditModalIsClosed(true)
-          }}
-        >
-          Update
-        </ModalUpdateButton>
-      </ModalInternalWrappingDiv>
-    </ModalBox>
-  )
-}
-
-const GifModalInternalWrappingDiv = styled.div`
-  display: grid;
-  justify-content: center;
-  align-content: center;
-  grid-gap: 10px;
-  font-size: 20px;
-`
-const GifCanvas = styled(P5Canvas)`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`
-
-const GifModalUpdateButton = styled.button`
-  display: block;
-  width: 80%;
-  height: 50px;
-  margin: 10px auto 0;
-  font-size: 18px;
-  border-radius: 5px;
-`
-
-const GenerateGifModal: React.FC = () => {
-  const [editModalIsClosed, setEditModalIsClosed] = useState(true)
-  const polygonContext = useContext(polygonGroupsStateContext)
-  const backgroundState = useContext(backgroundStateContext)
-  const [canUpdate, setCanUpdate] = useState(true)
-
-  return (
-    <ModalBox
-      buttonText="Share Gif"
-      title="Share Gif"
-      StyledButton={NavigationButton}
-      isClosed={editModalIsClosed}
-      setIsClosed={setEditModalIsClosed}
-    >
-      <ModalInternalWrappingDiv>
-        {!editModalIsClosed && (
-          <GifCanvas
-            sketch={generateGifSketch({
-              polygonGroups: polygonContext,
-              windowSize: { height: 250, width: 250 },
-              rgbaBackgroundColour: backgroundState.rgba,
-              rgbBackgroundColour: backgroundState.rgb,
-              shouldRedrawBackground: backgroundState.shouldRedraw,
-            })}
-          />
-        )}
-        <ModalUpdateButton
-          type="button"
-          disabled={!canUpdate}
-          onClick={() => {
-            // setCanUpdate(false)
-            setEditModalIsClosed(true)
-          }}
-        >
-          Update
-        </ModalUpdateButton>
-      </ModalInternalWrappingDiv>
-    </ModalBox>
-  )
-}
 
 const App: React.FC = () => {
   const navigationState = useContext(navigationStateContext)
