+++ src/components/__tests__/EditBackgroundModal.test.tsx
@@ -1,0 +1,6 @@
+import React from "react"
+import { render } from "@testing-library/react"
+
+describe("EditBackgroundModal Component", () => {
+  it.todo("should make some tests")
+})
