+++ src/components/__tests__/ShareGifModal.test.tsx
@@ -1,0 +1,6 @@
+import React from "react"
+import { render } from "@testing-library/react"
+
+describe("ShareGifModal Component", () => {
+  it.todo("should make some tests")
+})
